refactor: move post_notifier_provider.dart
