+++ lib/view/page/gallery/gallery_post_page.dart
@@ -8,8 +8,8 @@ import 'package:share_plus/share_plus.dart';
 import '../../../i18n/strings.g.dart';
 import '../../../model/account.dart';
 import '../../../provider/api/gallery_post_notifier_provider.dart';
-import '../../../provider/api/post_notifier_provider.dart';
 import '../../../provider/api/user_gallery_posts_notifier_provider.dart';
+import '../../../provider/post_notifier_provider.dart';
 import '../../../util/copy_text.dart';
 import '../../../util/future_with_dialog.dart';
 import '../../../util/launch_url.dart';
