+++ lib/view/page/page/page_page.dart
@@ -8,8 +8,8 @@ import 'package:share_plus/share_plus.dart';
 import '../../../i18n/strings.g.dart';
 import '../../../model/account.dart';
 import '../../../provider/api/page_provider.dart';
-import '../../../provider/api/post_notifier_provider.dart';
 import '../../../provider/api/user_pages_notifier_provider.dart';
+import '../../../provider/post_notifier_provider.dart';
 import '../../../util/copy_text.dart';
 import '../../../util/extract_url.dart';
 import '../../../util/future_with_dialog.dart';
