+++ lib/view/page/play/play_page.dart
@@ -7,8 +7,8 @@ import 'package:share_plus/share_plus.dart';
 import '../../../i18n/strings.g.dart';
 import '../../../model/account.dart';
 import '../../../provider/api/play_notifier_provider.dart';
-import '../../../provider/api/post_notifier_provider.dart';
 import '../../../provider/misskey_colors_provider.dart';
+import '../../../provider/post_notifier_provider.dart';
 import '../../../util/copy_text.dart';
 import '../../../util/launch_url.dart';
 import '../../widget/ad_widget.dart';
