+++ lib/view/page/tag/tag_page.dart
@@ -4,7 +4,7 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 import '../../../i18n/strings.g.dart';
 import '../../../model/account.dart';
-import '../../../provider/api/post_notifier_provider.dart';
+import '../../../provider/post_notifier_provider.dart';
 import 'tag_notes.dart';
 import 'tag_users.dart';
