+++ lib/view/page/timelines_page.dart
@@ -10,10 +10,10 @@ import '../../extension/scroll_controller_extension.dart';
 import '../../i18n/strings.g.dart';
 import '../../model/tab_type.dart';
 import '../../provider/api/i_notifier_provider.dart';
-import '../../provider/api/post_notifier_provider.dart';
 import '../../provider/emojis_notifier_provider.dart';
 import '../../provider/general_settings_notifier_provider.dart';
 import '../../provider/misskey_colors_provider.dart';
+import '../../provider/post_notifier_provider.dart';
 import '../../provider/streaming/main_stream_notifier_provider.dart';
 import '../../provider/timeline_scroll_controller_provider.dart';
 import '../../provider/timeline_tab_index_notifier_provider.dart';
