+++ lib/view/widget/note_footer.dart
@@ -12,11 +12,11 @@ import '../../model/account.dart';
 import '../../provider/account_settings_notifier_provider.dart';
 import '../../provider/api/i_notifier_provider.dart';
 import '../../provider/api/meta_notifier_provider.dart';
-import '../../provider/api/post_notifier_provider.dart';
 import '../../provider/appear_note_provider.dart';
 import '../../provider/general_settings_notifier_provider.dart';
 import '../../provider/note_provider.dart';
 import '../../provider/notes_notifier_provider.dart';
+import '../../provider/post_notifier_provider.dart';
 import '../../util/future_with_dialog.dart';
 import '../../util/launch_url.dart';
 import '../dialog/clip_dialog.dart';
