+++ lib/view/widget/play_widget.dart
@@ -14,10 +14,10 @@ import '../../provider/accounts_notifier_provider.dart';
 import '../../provider/aiscript_storage_notifier_provider.dart';
 import '../../provider/api/i_notifier_provider.dart';
 import '../../provider/api/play_notifier_provider.dart';
-import '../../provider/api/post_notifier_provider.dart';
 import '../../provider/dio_provider.dart';
 import '../../provider/emojis_notifier_provider.dart';
 import '../../provider/misskey_colors_provider.dart';
+import '../../provider/post_notifier_provider.dart';
 import '../../rust/api/aiscript.dart';
 import '../../rust/api/aiscript/api.dart';
 import '../../rust/api/aiscript/play.dart';
