+++ lib/view/widget/user_sheet.dart
@@ -10,8 +10,8 @@ import '../../provider/accounts_notifier_provider.dart';
 import '../../provider/api/i_notifier_provider.dart';
 import '../../provider/api/meta_notifier_provider.dart';
 import '../../provider/api/misskey_provider.dart';
-import '../../provider/api/post_notifier_provider.dart';
 import '../../provider/api/user_notifier_provider.dart';
+import '../../provider/post_notifier_provider.dart';
 import '../../util/copy_text.dart';
 import '../../util/future_with_dialog.dart';
 import '../../util/launch_url.dart';
